# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Used Books India API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  createdAt: string;
}

export interface CreateCategoryBody {
  name: string;
  adminPassword: string;
}

export type BookCondition = (typeof BookCondition)[keyof typeof BookCondition];

export const BookCondition = {
  like_new: "like_new",
  good: "good",
  fair: "fair",
  poor: "poor",
} as const;

export interface Book {
  id: number;
  title: string;
  author: string;
  /** @nullable */
  description: string | null;
  price: number;
  /** @nullable */
  originalPrice: number | null;
  condition: BookCondition;
  /** @nullable */
  imageUrl: string | null;
  categoryId: number;
  /** @nullable */
  categoryName: string | null;
  inStock: boolean;
  createdAt: string;
}

export type CreateBookBodyCondition =
  (typeof CreateBookBodyCondition)[keyof typeof CreateBookBodyCondition];

export const CreateBookBodyCondition = {
  like_new: "like_new",
  good: "good",
  fair: "fair",
  poor: "poor",
} as const;

export interface CreateBookBody {
  title: string;
  author: string;
  description?: string;
  price: number;
  originalPrice?: number;
  condition: CreateBookBodyCondition;
  imageUrl?: string;
  categoryId: number;
  inStock?: boolean;
  adminPassword: string;
}

export type UpdateBookBodyCondition =
  (typeof UpdateBookBodyCondition)[keyof typeof UpdateBookBodyCondition];

export const UpdateBookBodyCondition = {
  like_new: "like_new",
  good: "good",
  fair: "fair",
  poor: "poor",
} as const;

export interface UpdateBookBody {
  title?: string;
  author?: string;
  description?: string;
  price?: number;
  originalPrice?: number;
  condition?: UpdateBookBodyCondition;
  imageUrl?: string;
  categoryId?: number;
  inStock?: boolean;
  adminPassword: string;
}

export interface OrderItem {
  bookId: number;
  title: string;
  author: string;
  price: number;
  quantity: number;
}

export interface CreateOrderBody {
  customerName: string;
  email: string;
  phone: string;
  addressLine1: string;
  addressLine2?: string;
  city: string;
  state: string;
  pincode: string;
  items: OrderItem[];
  totalAmount: number;
  notes?: string;
}

export interface OrderConfirmation {
  success: boolean;
  message: string;
  orderId: string;
}

export interface AdminLoginBody {
  password: string;
}

export interface AdminLoginResponse {
  success: boolean;
  token: string;
}

export type ListBooksParams = {
  categoryId?: number;
  search?: string;
};