# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Used Books India API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all categories
 */
export const ListCategoriesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  slug: zod.string(),
  createdAt: zod.date(),
});
export const ListCategoriesResponse = zod.array(ListCategoriesResponseItem);

/**
 * @summary Create a category (admin)
 */
export const CreateCategoryBody = zod.object({
  name: zod.string(),
  adminPassword: zod.string(),
});

/**
 * @summary Delete a category (admin)
 */
export const DeleteCategoryParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List books with optional category filter
 */
export const ListBooksQueryParams = zod.object({
  categoryId: zod.coerce.number().optional(),
  search: zod.coerce.string().optional(),
});

export const ListBooksResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  author: zod.string(),
  description: zod.string().nullable(),
  price: zod.number(),
  originalPrice: zod.number().nullable(),
  condition: zod.enum(["like_new", "good", "fair", "poor"]),
  imageUrl: zod.string().nullable(),
  categoryId: zod.number(),
  categoryName: zod.string().nullable(),
  inStock: zod.boolean(),
  createdAt: zod.date(),
});
export const ListBooksResponse = zod.array(ListBooksResponseItem);

/**
 * @summary Add a book (admin)
 */
export const CreateBookBody = zod.object({
  title: zod.string(),
  author: zod.string(),
  description: zod.string().optional(),
  price: zod.number(),
  originalPrice: zod.number().optional(),
  condition: zod.enum(["like_new", "good", "fair", "poor"]),
  imageUrl: zod.string().optional(),
  categoryId: zod.number(),
  inStock: zod.boolean().optional(),
  adminPassword: zod.string(),
});

/**
 * @summary Get a single book
 */
export const GetBookParams = zod.object({
  id: zod.coerce.number(),
});

export const GetBookResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  author: zod.string(),
  description: zod.string().nullable(),
  price: zod.number(),
  originalPrice: zod.number().nullable(),
  condition: zod.enum(["like_new", "good", "fair", "poor"]),
  imageUrl: zod.string().nullable(),
  categoryId: zod.number(),
  categoryName: zod.string().nullable(),
  inStock: zod.boolean(),
  createdAt: zod.date(),
});

/**
 * @summary Update a book (admin)
 */
export const UpdateBookParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateBookBody = zod.object({
  title: zod.string().optional(),
  author: zod.string().optional(),
  description: zod.string().optional(),
  price: zod.number().optional(),
  originalPrice: zod.number().optional(),
  condition: zod.enum(["like_new", "good", "fair", "poor"]).optional(),
  imageUrl: zod.string().optional(),
  categoryId: zod.number().optional(),
  inStock: zod.boolean().optional(),
  adminPassword: zod.string(),
});

export const UpdateBookResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  author: zod.string(),
  description: zod.string().nullable(),
  price: zod.number(),
  originalPrice: zod.number().nullable(),
  condition: zod.enum(["like_new", "good", "fair", "poor"]),
  imageUrl: zod.string().nullable(),
  categoryId: zod.number(),
  categoryName: zod.string().nullable(),
  inStock: zod.boolean(),
  createdAt: zod.date(),
});

/**
 * @summary Delete a book (admin)
 */
export const DeleteBookParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Submit an order
 */
export const CreateOrderBody = zod.object({
  customerName: zod.string(),
  email: zod.string(),
  phone: zod.string(),
  addressLine1: zod.string(),
  addressLine2: zod.string().optional(),
  city: zod.string(),
  state: zod.string(),
  pincode: zod.string(),
  items: zod.array(
    zod.object({
      bookId: zod.number(),
      title: zod.string(),
      author: zod.string(),
      price: zod.number(),
      quantity: zod.number(),
    }),
  ),
  totalAmount: zod.number(),
  notes: zod.string().optional(),
});

/**
 * @summary Admin login
 */
export const AdminLoginBody = zod.object({
  password: zod.string(),
});

export const AdminLoginResponse = zod.object({
  success: zod.boolean(),
  token: zod.string(),
});