# lib/api-zod/src/generated/types/book.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Used Books India API
 * OpenAPI spec version: 0.1.0
 */
import type { BookCondition } from "./bookCondition";

export interface Book {
  id: number;
  title: string;
  author: string;
  /** @nullable */
  description: string | null;
  price: number;
  /** @nullable */
  originalPrice: number | null;
  condition: BookCondition;
  /** @nullable */
  imageUrl: string | null;
  categoryId: number;
  /** @nullable */
  categoryName: string | null;
  inStock: boolean;
  createdAt: Date;
}