# lib/api-zod/src/generated/types/createBookBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Used Books India API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateBookBodyCondition } from "./createBookBodyCondition";

export interface CreateBookBody {
  title: string;
  author: string;
  description?: string;
  price: number;
  originalPrice?: number;
  condition: CreateBookBodyCondition;
  imageUrl?: string;
  categoryId: number;
  inStock?: boolean;
  adminPassword: string;
}