# lib/api-zod/src/generated/types/createCategoryBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Used Books India API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCategoryBody {
  name: string;
  adminPassword: string;
}