# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Used Books India API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItem {
  bookId: number;
  title: string;
  author: string;
  price: number;
  quantity: number;
}