# lib/api-zod/src/generated/types/updateBookBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Used Books India API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateBookBodyCondition } from "./updateBookBodyCondition";

export interface UpdateBookBody {
  title?: string;
  author?: string;
  description?: string;
  price?: number;
  originalPrice?: number;
  condition?: UpdateBookBodyCondition;
  imageUrl?: string;
  categoryId?: number;
  inStock?: boolean;
  adminPassword: string;
}